Add some (temporary) test code for *out* and *err*
These dynamic vars get rebound by nrepl. So when we want to output
something on server side (to server console/log) when within an nRepl
session, we need to bind *out* and *err* to their original values.

This is mainly test-code, keep for now for also testing logging, which
seems quote complicated.