Make more robust with UncaughtExceptionHandler and keep-alive-loop
See crashes without logs when using raynes/conch for starting
shell-processes. Conch uses Threads and Futures for this in the
high-level functions. With low-level functions, there seem to be less
issues. Need to get experience. Could restart nrepl when it crashes,
and could also restart the JVM from the shell script when it crashes.